+++ base/include/rtai_mbx.h
@@ -156,9 +156,9 @@ int rt_named_mbx_delete(struct rt_mailbox *mbx);
 extern "C" {
 #endif /* __cplusplus */
 
-RTAI_PROTO(void *,rt_typed_mbx_init,(int name, int size, int qtype))
+RTAI_PROTO(void *, rt_typed_mbx_init, (unsigned long name, int size, int qtype))
 {
-	struct { int name; int size; int qtype; } arg = { name, size, qtype };
+	struct { unsigned long name; int size; int qtype; } arg = { name, size, qtype };
 	return rtai_lxrt(BIDX, SIZARG, LXRT_MBX_INIT, &arg).v[LOW];
 }
 
@@ -201,73 +201,73 @@ RTAI_PROTO(int, rt_named_mbx_delete,(struct rt_mailbox *mbx))
 
 RTAI_PROTO(int, rt_mbx_send,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_SEND, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_send_wp,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_SEND_WP, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_send_if,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_SEND_IF, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_send_until,(struct rt_mailbox *mbx, void *msg, int msg_size, RTIME time))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; RTIME time; int space; } arg = { mbx, (char *)msg, msg_size, time, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; RTIME time; long space; } arg = { mbx, (char *)msg, msg_size, time, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_SEND_UNTIL, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_send_timed,(struct rt_mailbox *mbx, void *msg, int msg_size, RTIME delay))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; RTIME delay; int space; } arg = { mbx, (char *)msg, msg_size, delay, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; RTIME delay; long space; } arg = { mbx, (char *)msg, msg_size, delay, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_SEND_TIMED, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_ovrwr_send,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_OVRWR_SEND, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_evdrp,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_EVDRP, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_receive,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_RECEIVE, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_receive_wp,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_RECEIVE_WP, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_receive_if,(struct rt_mailbox *mbx, void *msg, int msg_size))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; int space; } arg = { mbx, (char *)msg, msg_size, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; long space; } arg = { mbx, (char *)msg, msg_size, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_RECEIVE_IF, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_receive_until,(struct rt_mailbox *mbx, void *msg, int msg_size, RTIME time))
 {
-	struct { struct rt_mailbox *mbx; void *msg; int msg_size; RTIME time; int space; } arg = { mbx, (char *)msg, msg_size, time, 0 };
+	struct { struct rt_mailbox *mbx; void *msg; long msg_size; RTIME time; long space; } arg = { mbx, (char *)msg, msg_size, time, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_RECEIVE_UNTIL, &arg).i[LOW];
 }
 
 RTAI_PROTO(int, rt_mbx_receive_timed,(struct rt_mailbox *mbx, void *msg, int msg_size, RTIME delay))
 {
-	struct { struct rt_mailbox *mbx; char *msg; int msg_size; RTIME delay; int space; } arg = { mbx, (char *)msg, msg_size, delay, 0 };
+	struct { struct rt_mailbox *mbx; char *msg; long msg_size; RTIME delay; long space; } arg = { mbx, (char *)msg, msg_size, delay, 0 };
 	return (int)rtai_lxrt(BIDX, SIZARG, MBX_RECEIVE_TIMED, &arg).i[LOW];
 }
